[nextest] move config support out into another crate
Will build out a separate tool that uses the same config in an upcoming
PR.
