+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
     "datatest-stable",
+    "nextest/config",
     "nextest/runner",
     "quick-junit",
 ]
