+++ nextest/config/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "nextest-config"
+version = "0.1.0"
+authors = ["Diem Association <opensource@diem.com>"]
+license = "Apache-2.0 OR MIT"
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0.40"
+camino = "1.0.4"
+serde = { version = "1.0.125", features = ["derive"] }
+toml = "0.5.8"
