+++ nextest/config/src/config.rs
@@ -1,24 +1,16 @@
 // Copyright (c) The diem-devtools Contributors
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-//! Manages configuration for nextest.
+//! Configuration for nextest.
 
-use crate::reporter::FailureOutput;
+use crate::errors::*;
 use anyhow::Context;
 use camino::{Utf8Path, Utf8PathBuf};
 use serde::Deserialize;
 use std::{
     collections::{hash_map::Entry, HashMap},
-    error, fmt, io,
+    fmt,
 };
-use structopt::StructOpt;
-
-#[derive(Debug, StructOpt)]
-pub struct ConfigOpts {
-    /// Config file [default: <workspace-root>/nextest.toml]
-    #[structopt(long)]
-    config_file: Option<Utf8PathBuf>,
-}
 
 /// Configuration for nextest.
 #[derive(Clone, Debug, Deserialize)]
@@ -71,11 +63,11 @@ Custom, repository-specific configuration is layered on top of the default confi
     ///
     /// If the file isn't specified and the directory doesn't have `nextest.toml`, uses the default
     /// config options.
-    pub fn from_opts(
-        config_opts: &ConfigOpts,
+    pub fn from_sources(
+        config_file: Option<&Utf8Path>,
         workspace_root: &Utf8Path,
     ) -> Result<Self, ConfigReadError> {
-        let config = Self::read_from_sources(config_opts.config_file.as_deref(), workspace_root)?;
+        let config = Self::read_from_sources(config_file, workspace_root)?;
         config.validate()?;
         Ok(config)
     }
@@ -219,22 +211,22 @@ impl<'cfg> NextestProfile<'cfg> {
     }
 
     /// Returns the name of this test run.
-    pub(crate) fn name(&self) -> &'cfg str {
+    pub fn name(&self) -> &'cfg str {
         &self.inner.name
     }
 
     /// Returns the number of retries.
-    pub(crate) fn retries(&self) -> usize {
+    pub fn retries(&self) -> usize {
         self.inner.retries
     }
 
     /// Returns the metadata configuration.
-    pub(crate) fn metadata_config(&self) -> &'cfg MetadataConfig {
+    pub fn metadata_config(&self) -> &'cfg MetadataConfig {
         self.metadata
     }
 
     /// Returns the configuration for the situations in which failure is output.
-    pub(crate) fn failure_output(&self) -> FailureOutput {
+    pub fn failure_output(&self) -> FailureOutput {
         self.inner.failure_output
     }
 }
@@ -262,195 +254,64 @@ struct NextestProfileImpl {
 /// Metadata configuration for nextest.
 #[derive(Clone, Debug, Deserialize)]
 #[serde(rename_all = "kebab-case")]
-pub(crate) struct MetadataConfig {
+pub struct MetadataConfig {
     /// The directory where metadata files are stored. This is relative to the workspace root.
-    pub(crate) dir: Utf8PathBuf,
+    pub dir: Utf8PathBuf,
 
     /// Output metadata in the JUnit format in addition to the canonical format.
     #[serde(default)]
-    pub(crate) junit: Option<Utf8PathBuf>,
+    pub junit: Option<Utf8PathBuf>,
 }
 
-/// An error that occurred while reading the config.
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum ConfigReadError {
-    /// An error occurred while reading the file.
-    Read { file: Utf8PathBuf, err: io::Error },
-
-    /// An error occurred while deserializing the file into TOML.
-    Toml {
-        file: Utf8PathBuf,
-        err: toml::de::Error,
-    },
-
-    /// The default profile wasn't found.
-    DefaultProfileNotFound {
-        default_profile: String,
-        all_profiles: Vec<String>,
-    },
-
-    /// The metadata key wasn't found for a profile.
-    MetadataKeyNotFound {
-        profile: String,
-        metadata_key: String,
-        all_keys: Vec<String>,
-    },
-
-    /// The repository config defines a profile or metadata key that's already part of the default
-    /// config.
-    DuplicateKey {
-        /// The file which contains the key that's duplicated.
-        file: Utf8PathBuf,
-
-        /// The key that's duplicated.
-        key: String,
-
-        /// The kind of key that's duplicate: either "profile" or "metadata" currently.
-        kind: &'static str,
-    },
+#[derive(Copy, Clone, Debug, Eq, PartialEq, Deserialize)]
+#[serde(rename_all = "kebab-case")]
+pub enum FailureOutput {
+    Immediate,
+    ImmediateFinal,
+    Final,
+    Never,
 }
 
-impl ConfigReadError {
-    fn read(file: impl Into<Utf8PathBuf>, err: io::Error) -> Self {
-        ConfigReadError::Read {
-            file: file.into(),
-            err,
-        }
+impl FailureOutput {
+    pub fn variants() -> [&'static str; 4] {
+        ["immediate", "immediate-final", "final", "never"]
     }
 
-    fn toml(file: impl Into<Utf8PathBuf>, err: toml::de::Error) -> Self {
-        ConfigReadError::Toml {
-            file: file.into(),
-            err,
-        }
-    }
-
-    fn default_profile_not_found(
-        default_profile: impl Into<String>,
-        all_profiles: impl IntoIterator<Item = impl Into<String>>,
-    ) -> Self {
-        let mut all_profiles: Vec<_> = all_profiles.into_iter().map(|s| s.into()).collect();
-        all_profiles.sort_unstable();
-        ConfigReadError::DefaultProfileNotFound {
-            default_profile: default_profile.into(),
-            all_profiles,
-        }
-    }
-
-    fn metadata_key_not_found(
-        profile: impl Into<String>,
-        metadata_key: impl Into<String>,
-        all_keys: impl IntoIterator<Item = impl Into<String>>,
-    ) -> Self {
-        let mut all_keys: Vec<_> = all_keys.into_iter().map(|s| s.into()).collect();
-        all_keys.sort_unstable();
-        ConfigReadError::MetadataKeyNotFound {
-            profile: profile.into(),
-            metadata_key: metadata_key.into(),
-            all_keys,
-        }
-    }
-
-    fn duplicate_key(
-        file: impl Into<Utf8PathBuf>,
-        key: impl Into<String>,
-        kind: &'static str,
-    ) -> Self {
-        ConfigReadError::DuplicateKey {
-            file: file.into(),
-            key: key.into(),
-            kind,
-        }
-    }
-}
-
-impl fmt::Display for ConfigReadError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    pub fn is_immediate(self) -> bool {
         match self {
-            ConfigReadError::Read { file, .. } => write!(f, "failed to read file {}", file),
-            ConfigReadError::Toml { file, .. } => {
-                write!(f, "failed to deserialize TOML from file {}", file)
-            }
-            ConfigReadError::DefaultProfileNotFound {
-                default_profile,
-                all_profiles,
-            } => {
-                write!(
-                    f,
-                    "default profile '{}' not found (known profiles: {})",
-                    default_profile,
-                    all_profiles.join(", ")
-                )
-            }
-            ConfigReadError::MetadataKeyNotFound {
-                profile,
-                metadata_key,
-                all_keys,
-            } => write!(
-                f,
-                "for profile '{}', metadata key '{}' not found (known keys: {})",
-                profile,
-                metadata_key,
-                all_keys.join(", ")
-            ),
-            ConfigReadError::DuplicateKey { file, key, kind } => write!(
-                f,
-                "in {}, {} key '{}' duplicated from default config",
-                file, kind, key
-            ),
+            FailureOutput::Immediate | FailureOutput::ImmediateFinal => true,
+            FailureOutput::Final | FailureOutput::Never => false,
         }
     }
-}
 
-impl error::Error for ConfigReadError {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
+    pub fn is_final(self) -> bool {
         match self {
-            ConfigReadError::Read { err, .. } => Some(err),
-            ConfigReadError::Toml { err, .. } => Some(err),
-            ConfigReadError::DefaultProfileNotFound { .. }
-            | ConfigReadError::MetadataKeyNotFound { .. }
-            | ConfigReadError::DuplicateKey { .. } => None,
+            FailureOutput::Final | FailureOutput::ImmediateFinal => true,
+            FailureOutput::Immediate | FailureOutput::Never => false,
         }
     }
 }
 
-#[derive(Clone, Debug)]
-pub struct ProfileNotFound {
-    profile: String,
-    all_profiles: Vec<String>,
-}
-
-impl ProfileNotFound {
-    fn new(
-        profile: impl Into<String>,
-        all_profiles: impl IntoIterator<Item = impl Into<String>>,
-    ) -> Self {
-        let mut all_profiles: Vec<_> = all_profiles.into_iter().map(|s| s.into()).collect();
-        all_profiles.sort_unstable();
-        Self {
-            profile: profile.into(),
-            all_profiles,
+impl fmt::Display for FailureOutput {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            FailureOutput::Immediate => write!(f, "immediate"),
+            FailureOutput::ImmediateFinal => write!(f, "immediate-final"),
+            FailureOutput::Final => write!(f, "final"),
+            FailureOutput::Never => write!(f, "never"),
         }
     }
 }
 
-impl fmt::Display for ProfileNotFound {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "profile '{}' not found (known profiles: {})",
-            self.profile,
-            self.all_profiles.join(", ")
-        )
+impl Default for FailureOutput {
+    fn default() -> Self {
+        FailureOutput::Immediate
     }
 }
-
-impl error::Error for ProfileNotFound {}
 
 #[cfg(test)]
 mod tests {
-    use crate::config::NextestConfig;
+    use super::*;
 
     #[test]
     fn default_config_is_valid() {
