+++ nextest/config/src/lib.rs
@@ -1,0 +1,11 @@
+// Copyright (c) The diem-devtools Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! Configuration for nextest.
+//!
+//! TODO: fill this out in more detail.
+
+mod config;
+pub mod errors;
+
+pub use config::*;
