+++ nextest/runner/Cargo.toml
@@ -21,7 +21,8 @@ serde_json = "1.0.63"
 signal-hook = "0.3.8"
 structopt = "0.3.21"
 termcolor = "1.1.2"
-toml = "0.5.8"
+
+nextest-config = { path = "../config" }
 quick-junit = { path = "../../quick-junit" }
 
 [dev-dependencies]
