+++ nextest/runner/src/lib.rs
@@ -1,7 +1,6 @@
 // Copyright (c) The diem-devtools Contributors
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-pub mod config;
 pub mod dispatch;
 mod metadata;
 pub mod output;
